pyproject.toml
[tool.poetry]
name = "cbp"
version = "0.0.1"
description = "Component Based Programming CLI"
authors = ["Peter Alexander"]
license = "MPLv2"
readme = "README.md"

[tool.poetry.scripts]
cbp = "cbp.main:app"

[tool.poetry.dependencies]
python = "^3.11"
typer = {extras = ["all"], version = "^0.9.0"}
pyzmq = "^25.1.1"
gevent = "^23.9.1"
strawberry-graphql = "^0.209.7"
sqlmodel = {git = "https://github.com/tiangolo/sqlmodel.git"}
python-graphql-client = "^0.4.3"
msgpack = "^1.0.7"


[tool.poetry.group.dev.dependencies]
ipython = "^8.16.1"
jupyterlab = "^4.0.6"
pytest = "^7.4.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
